Add files via upload
In the age of data-driven decision-making, the importance of Exploratory data analysis (EDA) cannot be overstated. EDA is the initial step in the data analysis process, where we aim to understand the data, its structure, and uncover hidden patterns or relationships that may not be immediately apparent. This project is a perfect opportunity to showcase the power of EDA in revealing valuable insights from raw data.